requirements.txt
# 核心深度学习框架 / Core Deep Learning Framework
# pip install --pre torch torchvision --index-url https://download.pytorch.org/whl/nightly/cu130
# pip install tensorrt --extra-index-url https://pypi.nvidia.com
ultralytics

# 计算机视觉与图像处理 / Computer Vision & Image Processing
opencv-python
numpy
Pillow

# MiDaS 模型依赖与实用工具 / MiDaS Model & Utilities
tqdm
requests
matplotlib
timm
onnx
